@@ -2,8 +2,9 @@ source 'https://rubygems.org'
 
 ruby '2.1.1'
 
+
 # Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.0.4'
+gem 'rails', '4.1.5'
 
 # Use postgresql as the database for Active Record
 gem 'pg'
@@ -62,5 +63,6 @@ gem 'axlsx'
 gem 'newrelic_rpm'
 gem 'rails-patch-json-encode'
 gem "oj"
+
 
 
